% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggre_mi.R
\name{aggre_mi}
\alias{aggre_mi}
\title{Aggregate linear regression results after multiple imputation}
\usage{
aggre_mi(betamap, stderrmat, n = 60, k = 3)
}
\arguments{
\item{betamap}{V x m matrix of the coefficient estimates (intercept, txt, covariates etc.) over m imputation}

\item{stderrmat}{V x m matrix of the standard error (intercept, txt, covariates etc.) over m imputation}

\item{n}{sample size}

\item{k}{number of predictors (group + covariates)}
}
\description{
Aggregate linear regression results after multiple imputation
}
\examples{
stderrmat=matrix(rnorm(30)^2,10,3)
betamap=matrix(rnorm(30),10,3)
aggre_mi(betamap = betamap, stderrmat = stderrmat)

}
