% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.cohend.array.R
\name{compute.cohend.array}
\alias{compute.cohend.array}
\title{Computing Cohen's d from MRI images}
\usage{
\method{compute}{cohend.array}(imgs1, imgs2, figure = TRUE,
  col = bluered(200), breaks = c(-100:100)/100 * 7)
}
\arguments{
\item{imgs1}{4d image of group 1}

\item{imgs2}{4d image of group 2}

\item{figure}{(default=TRUE) draw image}

\item{col}{(only if figure=TRUE, default=bluered(200))}

\item{breaks}{(only if figure=TRUE, default=c(-100:100)/100*7)}
}
\value{
cohen.dmap: image array of Cohen's d.
}
\description{
Title
}
\examples{
library(neurorct)
img.dim=c(10,12,15)
n=10
treatment=array(rnorm(img.dim[1]*img.dim[2]*img.dim[3]*n),dim=c(img.dim,n))
treatment[3:5,3:5,3:5,]  <- 10+treatment[3:5,3:5,3:5,]
control=array(rnorm(img.dim[1]*img.dim[2]*img.dim[3]*n),dim=c(img.dim,n))
control[3:5,3:5,3:5,]  <- 12+control[3:5,3:5,3:5,]
z=compute.cohend.array(imgs1=treatment,imgs2=control)
}
