% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.cohend.R
\name{compute.cohend}
\alias{compute.cohend}
\title{Computing Cohen's d from MRI images}
\usage{
compute.cohend(imgs1, imgs2, mask = array(1, dim = dim(imgs1)[1:3]),
  figure = TRUE, col = bluered(200), breaks = c(-100:100)/100 * 7)
}
\arguments{
\item{imgs1}{4d image of group 1}

\item{imgs2}{4d image of group 2}

\item{mask}{Mask file (default=array(1,dim=dim(imgs1)[1:3]))}

\item{figure}{(default=TRUE) draw image}

\item{col}{(only if figure=TRUE, default=bluered(200))}

\item{breaks}{(only if figure=TRUE, default=c(-100:100)/100*7)}
}
\value{
cohen.dmap: image array of Cohen's d.
}
\description{
Cohen's d computation for neuroimaging data
}
\details{
This function computes simple Cohen's d maps from neuroimaging data
for group level analysis. The formula we use is
\deqn{d=\frac{\bar{X}_1- \bar{X}_2}{s_{pooled}},}
\eqn{s_{pooled}} is the pooled
standard deviation in two groups.
}
\examples{
library(neurorct)
img.dim=c(10,12,15)
n=10
treatment=array(rnorm(img.dim[1]*img.dim[2]*img.dim[3]*n),dim=c(img.dim,n))
treatment[3:5,3:5,3:5,]  <- 10+treatment[3:5,3:5,3:5,]
control=array(rnorm(img.dim[1]*img.dim[2]*img.dim[3]*n),dim=c(img.dim,n))
control[3:5,3:5,3:5,]  <- 12+control[3:5,3:5,3:5,]
mask=array(0,img.dim);mask[2:8,2:11,2:14]<-1
z=compute.cohend(imgs1=treatment,imgs2=control,mask=mask)
}
