% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_r.R
\name{compute_r}
\alias{compute_r}
\title{Computing Cohen's d from MRI images}
\usage{
compute_r(imgs, x, mask = array(1, dim = dim(imgs1)[1:3]),
  figure = FALSE, col = bluered(200), breaks = c(-100:100)/100,
  method = "pearson")
}
\arguments{
\item{imgs}{4d images}

\item{x}{independent variables}

\item{mask}{Mask file (default=array(1,dim=dim(imgs1)[1:3]))}

\item{figure}{(default=TRUE) draw image}

\item{col}{(only if figure=TRUE, default=bluered(200))}

\item{breaks}{(only if figure=TRUE, default=c(-100:100)/100*7)}

\item{method}{(default='pearson') 'spearman', 'tau' can be used.}
}
\value{
rmap: image array of pearson's coefficients.
}
\description{
Pearson's r computation for neuroimaging data
}
\details{
This function computes voxel-wise pearson's correlation coefficients between
an independent variable and neuroimaging data.
}
\examples{
library(neurorct)
img.dim=c(10,12,15)
n=100
x=rnorm(n)
img=array(rnorm(img.dim[1]*img.dim[2]*img.dim[3]*n),dim=c(img.dim,n))
img[3:5,3:5,3:5,]  <- array(rep(x,each=27),c(3,3,3,n)) +img[3:5,3:5,3:5,]
mask=array(0,img.dim);mask[2:8,2:11,2:14]<-1
z=compute_r(imgs=img,x=x,mask=mask,figure=TRUE)
}
