% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_lm.R, R/mi_lmer.R
\name{fast_lm}
\alias{fast_lm}
\title{Fast linear regression}
\usage{
fast_lm(x, y)

fast_lm(x, y)
}
\arguments{
\item{x}{nxp input design matrix. It should not include intercept.}

\item{y}{qxn outcome matrix.}

\item{x}{nxp input design matrix. It should not include intercept.}

\item{y}{qxn outcome matrix.}
}
\value{
tmaps t-values for each coefficients

df Degree of freedome of each coefficients

tmaps t-values for each coefficients

df Degree of freedome of each coefficients
}
\description{
Fast lm

Multiple imputation with lmer
}
\details{
This function computes linear regression and computes t-scores for multivariate outcomes.

This function computes mixed effect regression after
voxel-wise multiple imputation.
}
\examples{
library(neurorct)
y=matrix(rnorm(100*1000),1000,100)
x=cbind(rnorm(100),rnorm(100))
system.time(a<-t(apply(y, 1, function(zz)summary(lm(zz~x))$coefficients[,3])))
system.time(a2<-fast_lm(x=x,y=y))
sum(abs(a-a2$tmap))

library(neurorct)
m.dim=c(20,30,40)
y.base=y.fu=array(0,c(dim(a),npergroup*2))
eff.d=0.8
npergroup=30
for (j in 1:(npergroup*2)){
set.seed(j)
y.base[,,,j]<-array(rnorm(m.dim[1]*m.dim[2]*m.dim[3]),dim=m.dim)
tmp2<-0.5*perturb*array(rnorm(m.dim[1]*m.dim[2]*m.dim[3]),dim=m.dim)
if (j<npergroup){tmp2[10:13,15:17,20:22]<-tmp2[10:13,15:17,20:22]+eff.d}
y.fu[,,,j]<-tmp2
}


}
