% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mask.R
\name{generate_mask}
\alias{generate_mask}
\title{Helper function: Generate mask nifti object}
\usage{
generate_mask(img, saveinfti = FALSE, file = "")
}
\arguments{
\item{img}{object of class nifti to generate mask}

\item{saveinfti}{if saveinfti = TRUE, save result to NIfTI file.}

\item{file}{the name or connection of the file where the csv file is saved to.}
}
\value{
object of class nifti as mask
}
\description{
Helper function: Generate mask nifti object
}
\examples{
library(neurorct)


}
