% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_y.R
\name{generate_y}
\alias{generate_y}
\title{Generate simulated nifti data examples}
\usage{
generate_y(n = 10)
}
\arguments{
\item{n}{sample size in each of the two groups. There are 2*n rows in the simulated data.}
}
\value{
an array
}
\description{
Generate simulated nifti data examples
}
