% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_image.R
\name{import_image}
\alias{import_image}
\title{Import imaging data for analysis}
\usage{
import_image(folderpath, maskpath = NULL, complete = TRUE,
  sub2 = TRUE)
}
\arguments{
\item{folderpath}{folder path that contains imaging files.}

\item{maskpath}{mask file.}

\item{complete}{if TRUE, only read in files for complete-case data analysis.}

\item{sub2}{if TRUE, subsample the image by factor of 2.}
}
\value{
a list of baseline matrix and follow-up matrix.
}
\description{
Import imaging data for analysis
}
