% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{Generate list of files for complete case analysis}
\usage{
list_files(folder)
}
\arguments{
\item{folder}{folder path that contains .img and .hdr files}
}
\value{
list of imaging files for complete case analysis
}
\description{
Generate list of files for complete case analysis
}
\examples{
library(neurorct)
myfolder = "./con_0001/"
list_files(folder = myfolder)

}
