% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_lmer.R
\name{mi_lmer}
\alias{mi_lmer}
\title{Multiple Imputation and running mixed effect regression}
\usage{
mi_lmer(x, y)
}
\arguments{
\item{x}{nxp input design matrix. It should not include intercept.}

\item{y}{qxn outcome matrix.}
}
\value{
tmaps t-values for each coefficients

df Degree of freedome of each coefficients
}
\description{
Multiple imputation with lmer
}
\details{
This function computes mixed effect regression after
voxel-wise multiple imputation.
}
\examples{
library(neurorct)
m.dim=c(20,30,40)
y.base=y.fu=array(0,c(dim(a),npergroup*2))
eff.d=0.8
npergroup=30
for (j in 1:(npergroup*2)){
set.seed(j)
y.base[,,,j]<-array(rnorm(m.dim[1]*m.dim[2]*m.dim[3]),dim=m.dim)
tmp2<-0.5*perturb*array(rnorm(m.dim[1]*m.dim[2]*m.dim[3]),dim=m.dim)
if (j<npergroup){tmp2[10:13,15:17,20:22]<-tmp2[10:13,15:17,20:22]+eff.d}
y.fu[,,,j]<-tmp2
}


}
