% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurorct_gui.R
\name{neurorct_gui}
\alias{neurorct_gui}
\title{Run the neurorct built-in shiny app}
\usage{
neurorct_gui()
}
\value{
run a Shiny app
}
\description{
Run the neurorct built-in shiny app
}
\examples{
library(neurorct)
neurorct_gui()


}
