% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Run the neurorct built-in shiny app}
\usage{
run_shiny()
}
\value{
run a Shiny app
}
\description{
Run the neurorct built-in shiny app
}
\examples{
library(neurorct)
run_shiny()


}
