% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_lm.R
\name{fast_lm}
\alias{fast_lm}
\title{Fast linear regression}
\usage{
fast_lm(x, y, bl = NULL, ncore = 5)
}
\arguments{
\item{x}{nxp input design matrix. It should not include intercept.}

\item{y}{qxn outcome matrix.}

\item{bl}{qxn baseline matrix.}
}
\value{
tmaps t-values for each coefficients

df Degree of freedome of each coefficients
}
\description{
Fast lm
}
\details{
This function computes linear regression and computes t-scores for multivariate outcomes.
}
\examples{
library(neurorct)
y=matrix(rnorm(100*1000),1000,100)
x=cbind(rnorm(100),rnorm(100))
system.time(a<-t(apply(y, 1, function(zz)summary(lm(zz~x))$coefficients[,3])))
system.time(a2<-fast_lm(x=x,y=y))
sum(abs(a-a2$tmap))

}
