% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mask.R
\name{generate_mask}
\alias{generate_mask}
\title{Helper function: Generate mask nifti object}
\usage{
generate_mask(img)
}
\arguments{
\item{nifti_object}{object of class nifti to generate mask}
}
\value{
object of class nifti as mask
}
\description{
Helper function: Generate mask nifti object
}
\examples{

}
