% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mask.R
\name{generate_mask}
\alias{generate_mask}
\title{Generate mask nifti object}
\usage{
generate_mask(nifti_object)
}
\arguments{
\item{nifti_object}{nifti template}
}
\value{
a mask nifti object
}
\description{
Generate mask nifti object
}
\examples{

}
