% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xmat.R
\name{generate_xmat}
\alias{generate_xmat}
\title{Generate simulated X matrix (i.e. design matrix excluding intercept)}
\usage{
generate_xmat(npergroup, type = "xs")
}
\arguments{
\item{npergroup}{sample size in each of the two groups}

\item{type}{specify study type. 'xs': Cross-sectional; 'long':longitudinal.}
}
\value{
a data.frame object of X matrix
}
\description{
Generate simulated X matrix (i.e. design matrix excluding intercept)
}
