% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_y.R
\name{generate_y}
\alias{generate_y}
\title{Generate simulated nifti data examples}
\usage{
generate_y(npergroup = 10, type = c("xs", "long"), saveinfti = FALSE,
  outfile = "")
}
\arguments{
\item{npergroup}{sample size in each of the two groups. There are 2*n rows in the simulated data.}

\item{type}{specify study type. 'xs': Cross-sectional; 'long':longitudinal.}

\item{saveinfti}{if saveinfti = TRUE, save result to NIfTI file.}

\item{outfile}{the name of the file where the NIfTI file is saved to.}
}
\value{
if cross-sectional, an array of dim c(64, 64, 32, 2*npergroup); if longitudinal, a list of 2 such arrays for basline and follow-up.
}
\description{
Generate simulated nifti data examples
}
\examples{
library(neurorct)
# cross-sectional data
xs_data = generate_y(35, type = 'xs')

# longitudinal data
long_data = generate_y(20, type = 'long')
bl = long_data$baseline
fu = long_data$follow_up


}
